Cleaner tag generation
- merge multiple attribute settings
- move id/class writing to the attar rendering so they can be changed further on
- allow unsetting classes by passing a class map with booleans

This is the first step towards changing html elements in situ, for repeated running of the template. That will allow generating the full html on the server and then using the same template on the client to do AJAX updates.
The idea is to change the rendering to create htmlElement objects which get children etc. The objects are retained and on a subsequent run the code will remove objects that don't match the current tag, create new ones or update existing ones.
This will allow mapping to arrays like d3.js does.